requirements.txt
certifi==2018.4.16
chardet==3.0.4
chromedriver==2.24.1
colorama==0.3.9
coverage==4.5.1
dataclasses==0.6
defusedxml==0.5.0
Django==2.0
django-allauth==0.36.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==2.0.0.dev1
django-forms-bootstrap==3.1.0
django-nose==1.4.5
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
djongo==1.2.28
gunicorn==19.8.1
idna==2.6
mongoengine==0.9.0
nose==1.3.7
oauthlib==2.1.0
pymongo==3.6.1
python-dotenv==0.8.2
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.12.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
